added check to image2md.py to check if README exists

+++ labguide/image2md.py
@@ -14,7 +14,12 @@ for i,file in enumerate(files):
     url_dict[i] = f'https://onstakinc.github.io/cisco-tetration-hol/labguide/{module}/images/{new_fn}'
 
 os.chdir('..')
-copy2('README.md', 'README.bak.md') 
+
+if os.path.exists('README.md'):
+    print("Backing up README.md")
+    copy2('README.md', 'README.bak.md') 
+else:
+    pass
 
 with open('README.md','w') as f:
     f.write('# Cisco Tetration - Hands-On Lab\n')
